requirements.txt
Django==4.2.3
djangorestframework==3.14.0

